move get cleaning method to config and add test
